requirements.txt
streamlit
pandas
python-dotenv
requests
pyodbc